requirements.txt
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
google-auth==2.10.0
google-auth-oauthlib==0.5.2
gspread==5.4.0
gspread-formatting==1.0.6
idna==3.3
iniconfig==1.1.1
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
slack-bolt==1.14.3
slack-sdk==3.18.1
tomli==2.0.1
urllib3==1.26.11
